backend/requirements.txt
Django==4.2.7
django-allauth==0.51.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dj-rest-auth==2.2.1
python-decouple==3.8
python-dotenv==1.1.1
whitenoise==6.9.0
gunicorn==23.0.0
mongoengine==0.29.1
pymongo==3.11.4
dnspython==2.6.1
google-api-python-client==2.179.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
youtube-transcript-api==0.6.2
openai==1.3.0
requests==2.32.4
yt-dlp==2023.12.30
youtube-comment-downloader==0.1.78
numpy
scikit-learn
sentence-transformers
qdrant-client
